Add a POSIX OS abstraction layer.
This targets the early POSIX editions and should be broadly compatible.

Unfortunately it therefore can't use any atomics that were imported into
POSIX from the ISO C standard recently. To still make the operations
actually atomic, a mutex is used to protect the operations. This is
obviously bad for performance.

Using the newer atomics, when available, is a possible next step.
Although it might then make more sense to just use a C11 OSAL for
everything.

Adding the OSAL sources to the fluidsynth binary is needed because it
uses fluid_test_file, which is not exported by the library. This is
problematic, as it will pull in other symbols into the binary as well
and needs to be reworked.